Minor restructure
